rename BgText to OverlayText and perform overall improvements (#33)
